Move select-workplace component into create-account folder
