+++ src/app/features/create-account/workplace/select-workplace/select-workplace.component.spec.ts
@@ -11,7 +11,7 @@ import { SharedModule } from '@shared/shared.module';
 import { fireEvent, render } from '@testing-library/angular';
 import { BehaviorSubject } from 'rxjs';
 
-import { RegistrationModule } from '../registration.module';
+import { RegistrationModule } from '../../../registration/registration.module';
 import { SelectWorkplaceComponent } from './select-workplace.component';
 
 describe('SelectWorkplaceComponent', () => {
